pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "daaily/version.py"

[tool.hatch.build.targets.wheel]
packages = ["daaily"]

[tool.hatch.build.targets.sdist]
exclude = [
  "/.github",
  "/samples",
  ".vscode",
  "/tests",
]

[project]
name = "daaily"
authors = [
  { name="Daaily Dev Team", email="dev@daaily.com" },
]
description = "Daaily API Library"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["httpx==0.27.0", "pydantic-settings==2.1.0"]

dynamic = ["version"]

[project.urls]
"Source code" = "https://github.com/DAAily/daailyapis-python-client"


[tool.ruff]
select = ["E", "F", "B", "Q", "C", "I"]
fixable = ["E", "F", "B", "Q", "C", "I"]
unfixable = ["F841","F401"]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "cms",
    "sheets",
]
per-file-ignores = {}
line-length = 88
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
target-version = "py312"

[tool.ruff.mccabe]
max-complexity = 10
